pyproject.toml
[project]
name = "hf-browser-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "browser-use>=0.1.40",
    "langchain>=0.3.21",
    "langchain-mcp-adapters>=0.0.6",
    "langgraph>=0.3.21",
    "lmnr[all]>=0.4.66",
    "python-dotenv>=1.0.1",
]
